% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnostics.R
\docType{package}
\name{fishnostics}
\alias{fishnostics}
\alias{fishnostics-package}
\title{fishnostics: Diagnostics for fisheries assessment models, specifically CASAL,
however, applicable to other fisheries assessments}
\description{
fishnostics: Diagnostics for fisheries assessment models, specifically CASAL,
however, applicable to other fisheries assessments
}

