% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot_ts_uncertainty}
\alias{plot_ts_uncertainty}
\title{Plot a timeseries with uncertainty}
\usage{
plot_ts_uncertainty(d, xs = NULL, d2 = NULL, probs = c(lwr = 0.025, mid =
  0.5, upr = 0.975), xlab = "", ylab = "", save = FALSE,
  file_name = NULL, file_type = "png", path = NULL,
  graph_pars = list(mfrow = c(1, 1), mar = c(4, 4, 1, 1), las = 2))
}
\arguments{
\item{d}{dataframe with columns representing each of the ys (could rename d to be the main series?)}

\item{xs}{vector of xs corresponding with the columns of d (if NULL, names of
d are used)}

\item{d2}{a dataframe with the same number of columns as d used to add a
second series as a line plot (same xs are used)}

\item{probs}{quantile probabilities for plotting}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{save}{logical to specify whether the plot should be saved (if TRUE a
file_name must be supplied)}

\item{file_name}{valid file name list to store model results}

\item{file_type}{default = png}

\item{graph_pars}{graphical parameters passed to par}
}
\description{
Plot a timeseries with uncertainty represented as a grey polygon with the
option of adding a second series to the plot
}

