% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_outputs.R
\name{process_two_series}
\alias{process_two_series}
\title{Process two Monte Carlo outputs}
\usage{
process_two_series(res, true_vars, est_vars, col_names = NULL, FUN)
}
\arguments{
\item{res}{dataframe of outputs from simulations}

\item{true_vars}{names of the columns that represent the true or known
parameters}

\item{est_vars}{names of columns that represent estimated parameters}

\item{col_names}{column names for the output (default=NULL will when
implemented generate these automatically)}

\item{FUN}{function to apply "rel_err" calculate the relative error}
}
\description{
Process two Monte Carlo outputs
}

