% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot}
\usage{
save_plot(file_name, file_type, path)
}
\arguments{
\item{file_name}{valid file name list to store model results passed to
\code{\link{save_plot}}}

\item{file_type}{default = png, passed to \code{\link{save_plot}}}
}
\description{
Save a plot
}

