% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnostics.R
\docType{package}
\name{fishnostics}
\alias{fishnostics}
\alias{fishnostics-package}
\title{fishnostics: Diagnostics for fisheries assessment models, specifically CASAL,
however, applicable to other fisheries assessments}
\description{
This package currently includes plots used to present model outputs along
with diagnostic plots.
}
\details{
The intention is that this package will be implemented in completely in
ggplot2, however, currently base and lattice are used.

Need to generate some example datsets to use for package examples
}

