% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnostics.R
\name{plotting}
\alias{plotting}
\title{Generic parameters for plotting}
\arguments{
\item{probs}{quantile probabilities for plotting}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{save}{logical to specify whether the plot should be saved (if TRUE a
file_name must be supplied)}

\item{file_name}{valid file name list to store model results}

\item{file_type}{default = png}

\item{graph_pars}{graphical parameters passed to par}
}
\description{
Generic parameters for plotting
}

